Refactor headers, change main class to "autoparser"
