+++ docs/autoparser.md
@@ -18,27 +18,27 @@ The unified auto-parser uses a pure differential, compositional approach (inspir
 
 ## Data Structures
 
-All structs are defined in [common/chat-diff-analyzer.h](common/chat-diff-analyzer.h).
+All structs are defined in [common/chat-auto-parser.h](common/chat-auto-parser.h).
 
 ### Top-Level: `autoparser` (main analyzer and generator)
 
-[common/chat-diff-analyzer.h:344-362](common/chat-diff-analyzer.h#L344-L362) — top-level analysis result aggregating `jinja_caps`, `reasoning`, `content`, and `tools` sub-analyses, plus `preserved_tokens` (union of all non-empty markers).
+[common/chat-auto-parser.h:367-388](common/chat-auto-parser.h#L367-L388) — top-level analysis result aggregating `jinja_caps`, `reasoning`, `content`, and `tools` sub-analyses, plus `preserved_tokens` (union of all non-empty markers).
 
 ### `analyze_reasoning`
 
-[common/chat-diff-analyzer.h:231-251](common/chat-diff-analyzer.h#L231-L251) — reasoning analysis result: `mode` enum, `start` marker (e.g. `<think>`), and `end` marker (e.g. `</think>`).
+[common/chat-auto-parser.h:254-274](common/chat-auto-parser.h#L254-L274) — reasoning analysis result: `mode` enum, `start` marker (e.g. `<think>`), and `end` marker (e.g. `</think>`).
 
 ### `analyze_content`
 
-[common/chat-diff-analyzer.h:257-272](common/chat-diff-analyzer.h#L257-L272) — content analysis result: `mode` enum, `start`/`end` markers, and `requires_nonnull_content` flag.
+[common/chat-auto-parser.h:280-295](common/chat-auto-parser.h#L280-L295) — content analysis result: `mode` enum, `start`/`end` markers, and `requires_nonnull_content` flag.
 
 ### `analyze_tools` and its sub-structs
 
-- [common/chat-diff-analyzer.h:153-171](common/chat-diff-analyzer.h#L153-L171) — `tool_format_analysis`: `mode` enum, `section_start/end`, `per_call_start/end`, JSON field names (`function_field`, `name_field`, `args_field`, `id_field`, `gen_id_field`), and format flags (`fun_name_is_key`, `tools_array_wrapped`, `uses_python_dicts`)
-- [common/chat-diff-analyzer.h:173-177](common/chat-diff-analyzer.h#L173-L177) — `tool_function_analysis`: `name_prefix`, `name_suffix`, `close` markers around function names
-- [common/chat-diff-analyzer.h:179-187](common/chat-diff-analyzer.h#L179-L187) — `tool_arguments_analysis`: `start/end` container markers, `name_prefix/suffix`, `value_prefix/suffix`, `separator`
-- [common/chat-diff-analyzer.h:189-194](common/chat-diff-analyzer.h#L189-L194) — `tool_id_analysis`: `pos` enum, `prefix`/`suffix` markers around call ID values
-- [common/chat-diff-analyzer.h:278-338](common/chat-diff-analyzer.h#L278-L338) — `analyze_tools`: aggregates the four sub-structs above
+- [common/chat-auto-parser.h:176-194](common/chat-auto-parser.h#L176-L194) — `tool_format_analysis`: `mode` enum, `section_start/end`, `per_call_start/end`, JSON field names (`function_field`, `name_field`, `args_field`, `id_field`, `gen_id_field`), and format flags (`fun_name_is_key`, `tools_array_wrapped`, `uses_python_dicts`)
+- [common/chat-auto-parser.h:196-200](common/chat-auto-parser.h#L196-L200) — `tool_function_analysis`: `name_prefix`, `name_suffix`, `close` markers around function names
+- [common/chat-auto-parser.h:202-210](common/chat-auto-parser.h#L202-L210) — `tool_arguments_analysis`: `start/end` container markers, `name_prefix/suffix`, `value_prefix/suffix`, `separator`
+- [common/chat-auto-parser.h:212-217](common/chat-auto-parser.h#L212-L217) — `tool_id_analysis`: `pos` enum, `prefix`/`suffix` markers around call ID values
+- [common/chat-auto-parser.h:301-361](common/chat-auto-parser.h#L301-L361) — `analyze_tools`: aggregates the four sub-structs above
 
 ### Enums
 
@@ -244,7 +244,7 @@ All analysis phases use the same factorized comparison function declared in [com
 compare_variants(tmpl, params_A, params_modifier)
 ```
 
-This creates variant B by applying a modifier lambda to a copy of `params_A`, renders both through the template, and computes a `diff_split` ([common/chat-diff-analyzer.h:28-36](common/chat-diff-analyzer.h#L28-L36)):
+This creates variant B by applying a modifier lambda to a copy of `params_A`, renders both through the template, and computes a `diff_split` ([common/chat-auto-parser.h:28-37](common/chat-auto-parser.h#L28-L37)):
 
 - `prefix` — common prefix between A and B
 - `suffix` — common suffix between A and B
@@ -341,7 +341,7 @@ Classification logic:
 
 ### Workarounds
 
-A workaround array in `chat-diff-analyzer.cpp` applies post-hoc patches after analysis. Each workaround is a lambda that inspects the template source and overrides analysis results. Current workarounds:
+A workaround array in `common/chat-diff-analyzer.cpp` applies post-hoc patches after analysis. Each workaround is a lambda that inspects the template source and overrides analysis results. Current workarounds:
 
 1. **Old Qwen/DeepSeek thinking templates** — source contains `content.split('</think>')`: sets `reasoning.mode = FORCED_OPEN` with `<think>`/`</think>` markers if no reasoning was detected
 2. **Granite 3.3** — source contains specific "Write your thoughts" text: forces `TAG_BASED` reasoning with `<think>`/`</think>` and `WRAPPED_WITH_REASONING` content with `<response>`/`</response>`
@@ -428,9 +428,8 @@ When `format.uses_python_dicts` is true (detected when single-quoted strings app
 
 | File                                      | Purpose                                                              |
 |-------------------------------------------|----------------------------------------------------------------------|
-| `common/chat-auto-parser.h`               | `peg_generator` class and `templates_params` struct                  |
+| `common/chat-auto-parser.h`               | All analysis structs, enums, `autoparser`, `peg_generator`, `templates_params` |
 | `common/chat-auto-parser-generator.cpp`   | Parser generator: `generate_parser()` and `build_parser()` methods   |
-| `common/chat-diff-analyzer.h`             | All analysis structs, enums, and sub-structs                         |
 | `common/chat-diff-analyzer.cpp`           | Differential analysis implementation and workarounds                 |
 | `common/chat-auto-parser-helpers.h/cpp`   | `calculate_diff_split()`, `segmentize_markers()`,                    |
 |                                           | `compare_variants()`, string helpers                                 |
@@ -512,7 +511,7 @@ The following templates have active tests in `tests/test-chat.cpp`:
 To support a new template format:
 
 1. **If it follows standard patterns** — The auto-parser should detect it automatically. Run `llama-debug-template-parser` to verify markers are correctly extracted.
-2. **If differential analysis extracts incorrect markers** — Add a workaround lambda to the `workarounds` vector in `chat-diff-analyzer.cpp`. Inspect the template source for a unique identifying substring.
+2. **If differential analysis extracts incorrect markers** — Add a workaround lambda to the `workarounds` vector in `common/chat-diff-analyzer.cpp`. Inspect the template source for a unique identifying substring.
 3. **If it needs fundamentally different handling** — Add a dedicated handler function in `chat.cpp` before the auto-parser block (as done for GPT-OSS, Functionary v3.2, and Ministral).
 
 ## Edge Cases and Quirks
