+++ tools/parser/debug-template-parser.cpp
@@ -1,6 +1,5 @@
 #include "../src/llama-grammar.h"
 #include "chat-auto-parser.h"
-#include "chat-diff-analyzer.h"
 #include "chat.h"
 #include "common.h"
 #include "gguf.h"
@@ -392,7 +391,8 @@ int main(int argc, char ** argv) {
             LOG_ERR("                           TEMPLATE ANALYSIS\n");
             LOG_ERR("================================================================================\n");
 
-            autoparser::autoparser analysis(chat_template);
+            autoparser::autoparser analysis;
+            analysis.analyze_template(chat_template);
 
             // Generate Parser
             autoparser::templates_params params;
